Moved Softmax method to HiddenNeuron and added Neuron Unit Tests
