+++ NeuralNetwork/HiddenNeuron.cs
@@ -1,6 +1,8 @@
-﻿namespace TradingBot
+﻿using System;
+
+namespace TradingBot
 {
-	internal class HiddenNeuron : Neuron
+	public class HiddenNeuron : Neuron
 	{
 		public double value;
 
@@ -8,6 +10,32 @@
 		{
 			this.Layer = layer;
 			this.Height = height;
+		}
+
+		public static double[] Softmax(double[] z)
+		{
+			double[] zExp = new double[z.Length];
+
+			for (int i = 0; i < zExp.Length; i++)
+			{
+				zExp[i] = Math.Exp(z[i]);
+			}
+
+			double sumZExp = 0;
+
+			for (int i = 0; i < zExp.Length; i++)
+			{
+				sumZExp += zExp[i];
+			}
+
+			double[] softmax = new double[z.Length];
+
+			for (int i = 0; i < zExp.Length; i++)
+			{
+				softmax[i] = zExp[i] / sumZExp;
+			}
+
+			return softmax;
 		}
 	}
 }
