+++ NeuralNetwork/NeuralNetwork.csproj
@@ -40,8 +40,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="HiddenNeuron.cs" />
     <Compile Include="InputData.cs" />
+    <Compile Include="InputNeuron.cs" />
     <Compile Include="NeuralNetwork.cs" />
+    <Compile Include="Neuron.cs" />
+    <Compile Include="OutputNeuron.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Synapse.cs" />
   </ItemGroup>
