+++ NeuralNetworkTests/NeuralNetworkTests.cs
@@ -11,20 +11,6 @@ namespace TradingBot.Tests
 	[TestClass()]
 	public class NeuralNetworkTests
 	{
-		[TestMethod()]
-		public void SoftmaxTest1()
-		{
-			double[] input = {1.0, 2.0, 3.0, 4.0, 1.0, 2.0, 3.0};
-			double[] expectedOutput = {0.024, 0.064, 0.175, 0.475, 0.024, 0.064, 0.175};
 
-			double[] output = NeuralNetwork.Softmax(input);
-
-			for (int i = 0; i < output.Length; i++)
-			{
-				output[i] = Math.Round(output[i], 3);
-
-				Assert.AreEqual(output[i], expectedOutput[i]);
-			}
-		}
 	}
 }
