+++ NeuralNetworkTests/NeuralNetworkTests.csproj
@@ -59,6 +59,7 @@
   <ItemGroup>
     <Compile Include="InputDataTests.cs" />
     <Compile Include="NeuralNetworkTests.cs" />
+    <Compile Include="NeuronTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
